% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_loadings.R
\name{center.truncate}
\alias{center.truncate}
\title{Trancate a string to no longer than 43 characters.
Where a string is longer than 43 characters, the first and last 20 characters sandwich three periods.
This can be useful for shortening long feature names before plotting.}
\usage{
center.truncate(x)
}
\arguments{
\item{x}{A string to truncate.}
}
\value{
A string where each has been reduced to 43 characters or left as is.
}
\description{
Trancate a string to no longer than 43 characters.
Where a string is longer than 43 characters, the first and last 20 characters sandwich three periods.
This can be useful for shortening long feature names before plotting.
}
\examples{
\dontrun{
center.truncate('A string to truncate if long enough')
}
}
