% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_analysis.R
\name{.nulldev}
\alias{.nulldev}
\title{A null device for running plot functions without outputting graphics.}
\usage{
.nulldev(file = nullfile(), ...)
}
\description{
A null device for running plot functions without outputting graphics.
}
