% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_analysis.R
\name{find.feature.associations}
\alias{find.feature.associations}
\title{Extract feature vs. feature association (mutual information) data for a multi-omics mixOmics (DIABLO) mode, and
plot this as a heatmap. This is the same data used to create the circos and network plots, but includes all
interactions for the top twenty features from each block.}
\usage{
find.feature.associations(diablo.tuned, block.count)
}
\arguments{
\item{block.count}{Number of blocks used in the model}

\item{diablo}{Trained mixOmics multi-omics (DIABLO) model}
}
\value{
Matrix including the association of all selected features across all blocks. E.g. for a three block
multi-omic model, this will include the top 20 features from each block, and so return a 60x60 covariance matrix
}
\description{
Extract feature vs. feature association (mutual information) data for a multi-omics mixOmics (DIABLO) mode, and
plot this as a heatmap. This is the same data used to create the circos and network plots, but includes all
interactions for the top twenty features from each block.
}
\examples{
\dontrun{
find.feature.associations(diablo.trained.analysis, 3)
}
}
