% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MixMC.R
\name{normalise.clr.within.features}
\alias{normalise.clr.within.features}
\title{Apply centred log-ratio (CLR) normalisation to other compositional data, but restrict normalisation to *within* features
only. This is another method for converting the compositional data, i.e. proportional data in the range 0..1 to
Euclidean space which is most appropriate for the linear models. The implementation is the same as CLR, but on
transposed input data (which is then transposed back to the input orientation). Note this is experimental, though it
does give a sensible normalisation.}
\usage{
normalise.clr.within.features(input, offset = 0)
}
\arguments{
\item{input}{Data as proportion 0-1}

\item{offset}{Optional offset to apply to raw data to avoid logging of zero values. Only needed if any zeroes are present - should generally be set very small, e.g. 0.000001.}
}
\value{
Data normalised by the within-feature CLR method
}
\description{
Apply centred log-ratio (CLR) normalisation to other compositional data, but restrict normalisation to *within* features
only. This is another method for converting the compositional data, i.e. proportional data in the range 0..1 to
Euclidean space which is most appropriate for the linear models. The implementation is the same as CLR, but on
transposed input data (which is then transposed back to the input orientation). Note this is experimental, though it
does give a sensible normalisation.
}
\examples{
\dontrun{
normalise.clr.within.features(data.proportional)
}
}
