% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna_utilities.R
\name{wgcna.determine.module.eigengenes}
\alias{wgcna.determine.module.eigengenes}
\title{Identify eigengenes that represent each module of genes.}
\usage{
wgcna.determine.module.eigengenes(expr.data, module.colors)
}
\arguments{
\item{expr.data}{Filtered and normalised expression data for genes to calculate eigengenes for.}

\item{module.colors}{A vector of color names identifying modules of genes, the length of which matches the number of columns in `expr.data`.}
}
\value{
The eigengenes for each module, one per input sample.
}
\description{
Identify eigengenes that represent each module of genes.
}
